pyproject.toml
[project]
name = "image_similarity"
version = "0.0.1"
description = "Image loading, preprocessing and clustering"
authors = [
    {name = "Jeronimo Gabande",email = "jedevar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "h5py (>=3.14.0,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
